requirements.txt
# CAPTCHA Solver Bot Dependencies

# Core dependencies
numpy>=1.21.0
opencv-python>=4.5.0
Pillow>=8.3.0
requests>=2.25.0
aiohttp>=3.7.0
asyncio-throttle>=1.0.0

# Computer Vision & OCR
pytesseract>=0.3.8
easyocr>=1.6.0
paddleocr>=2.6.0

# Machine Learning
torch>=1.9.0
torchvision>=0.10.0
tensorflow>=2.6.0
scikit-learn>=0.24.0
scikit-image>=0.18.0
onnx>=1.10.0
onnxruntime>=1.8.0

# Audio processing
librosa>=0.8.0
soundfile>=0.10.0
pydub>=0.25.0
speech-recognition>=3.8.0
whisper>=1.0.0

# Web automation
selenium>=4.0.0
webdriver-manager>=3.8.0
undetected-chromedriver>=3.4.0
playwright>=1.20.0
requests-html>=0.10.0

# Image processing
imagehash>=4.2.0
matplotlib>=3.4.0
seaborn>=0.11.0

# Data handling
pandas>=1.3.0
pyyaml>=5.4.0
jsonschema>=3.2.0
python-dotenv>=0.19.0

# HTTP clients
httpx>=0.24.0
cloudscraper>=1.2.60

# Utility libraries
tqdm>=4.62.0
colorama>=0.4.4
click>=8.0.0
rich>=10.0.0

# Testing
pytest>=6.2.0
pytest-asyncio>=0.15.0
pytest-mock>=3.6.0
pytest-cov>=2.12.0

# Code quality
black>=21.0.0
flake8>=3.9.0
mypy>=0.910
isort>=5.9.0

# Documentation
sphinx>=4.1.0
sphinx-rtd-theme>=0.5.0

# Logging and monitoring
structlog>=21.1.0
sentry-sdk>=1.3.0

# Performance
memory-profiler>=0.58.0
line-profiler>=3.3.0
psutil>=5.8.0

# Database (optional)
sqlite3>=3.35.0
sqlalchemy>=1.4.0

# GUI (optional)
tkinter>=8.6.0
customtkinter>=4.0.0

# API server (optional)
fastapi>=0.68.0
uvicorn>=0.15.0
pydantic>=1.8.0

# Development tools
ipython>=7.25.0
jupyter>=1.0.0
notebook>=6.4.0